# Do not try to calculate hitpoint for elements outside of the screen rect