This is one of these cases where everything good should be simple. 
**self.fb_hitPoint** call is expensive and there is no point to call it if we already know that rectIntersects verification failed