@qmfrederik 2017-05-23
@mykola-mokhnach Did you run perf tests on this? Here's what I get (using 10 requests):

* Current master: 781 ms +/- 10ms
* PR 597: 763 ms +/- 14ms

So about a 3% improvement in performance.

The gist of this is that `fb_hitPoint` is expensive, but `visibleFrame` is expensive, too. Both seem to re-use some cached variables; so if you call `visibleFrame` first, `fb_hitPoint` is fast(er).

Looking at the code, I was wondering if the following is true:

> If `self.frame` is entirely within `app.frame`, then `self.visibleFrame` and `self.fb_hitPoint` are also in `app.frame`

and

> If `self.frame` is entirely outside of `app.frame`, then `self.VisibleFrame` and `self.fb_hitPoint` are also outside of `app.Frame`

If that is true, you can avoid calling `visibleFrame` and `fb_hitPoint` in a lot of scenarios, and then I get the following results: 

* 552 ms +/- 7 ms

so about 30% improvement in performance.