@mykola-mokhnach 2017-05-24
thanks for the hint @qmfrederik 
I have replaced **visibilityFrame** property with **frame** in the condition. It's clear that if element's **frame** is outside of the visible viewport then **visibleFrame** will be outside of it as well.
Can you please the perf now?