@qmfrederik 2017-05-24
@mykola-mokhnach I re-ran the tests, but the results are pretty much the same. I think there is already some caching involved so it's the first call to hitpoint / visible frames which really cause the perf hit.

I think the fix would be to figure out if there's another way to anticipate the hitpoint will be either valid (inside the screen area) or invalid ( such as -1,-1). I'm guessing there must be another property which can give this information, but I didn't find one.

I'll probably keep looking at this the coming days and keep you updated.

PS:
- To measure the HTTP response time, I just load the /source URL with Chrome and use the Chrome developer tools
- To figure out which code blocks are taking time, I use `CFTimeInterval`:

```
(...)

#import <QuartzCore/QuartzCore.h>

(...)

  CFTimeInterval startTime = CACurrentMediaTime();

  if (CGRectIsEmpty(self.frame)) {
    return NO;
  }
  CFTimeInterval evaluateFrameEmptyTime = CACurrentMediaTime();

(...)


  NSLog(@"frameEmpty: %f, visibleFrame: %f, configuration: %f, frameIntersect: %f, hitPoint: %f, visible: %d, x: %f, y: %f, width: %f, height: %f, hitPoint: (%f, %f)",
        evaluateFrameEmptyTime - startTime,

(...)

```